Clean up Stories to be more readable, put more logic inside the stories themselves so you can see it on the Storybook page
