+++ stories/StatefulReactView/InternalState/InternalState.stories.js
@@ -8,25 +8,23 @@ import reducers from './reducers'
 import InternalState from './InternalState.md'
 import withReadme from 'storybook-readme/with-readme'
 import makeListModel from '../../utils/makeListModel'
-import Marionette from 'backbone.marionette'
 import StatefulReactView from '../../../src/hoc/StatefulReactView'
 
 const stories = storiesOf('StatefulReactView', module)
-
-const TodoAppView = StatefulReactView(Marionette.ItemView).extend({
-  template() {
-    return <TodoApp />
-  },
-})
-
 stories.add(
   'Internal State',
   withReadme(InternalState, () => {
+    const TodoAppView = StatefulReactView().extend({
+      template() {
+        return <TodoApp />
+      },
+    })
+
     const preloadedState = {
       todos: TaskSpec.generate(3),
       meta: makeListModel(),
     }
     const store = createStore({preloadedState, reducers})
-    return <ViewWrapper renderView={() => new TodoAppView({ store })} />
+    return <ViewWrapper renderView={() => new TodoAppView({store})} />
   }),
 )
